191031 Django RESTful

+++ django/crud_review/students/views.py
@@ -18,37 +18,74 @@ def detail(request, pk):
     return render(request, 'students/detail.html', context)
 
 def delete(request,pk):
-    student = Students.objects.get(pk=pk)
-    student.delete()
+
+    if request.method == 'POST':
+        student = Students.objects.get(pk=pk)
+        student.delete()
     
-    return redirect('/students/')
+        return redirect('students:index')
 
 def new(request):
-    return render(request, 'students/new.html')
+    if request.method == 'POST':
+        # students를 생성함.
+        name = request.POST.get('name')
+        age = request.POST.get('age')
 
-def create(request):
-    name = request.POST.get('name')
-    age = request.POST.get('age')
+        students = Students.objects.create(name= name , age = age)
+        students.save()
 
-    students = Students.objects.create(name= name , age = age)
+        return redirect('/students/')
+
+    else:        
+        # new page를 보여주면 됨.
+        return render(request, 'students/new.html')
+
     
-    return redirect('/students/')
+
+# def create(request):
+#     name = request.POST.get('name')
+#     age = request.POST.get('age')
+
+#     students = Students.objects.create(name= name , age = age)
+#     students.save()
+
+#     return redirect('/students/')
 
 def edit(request, pk):
     students = Students.objects.get(pk=pk)
-    context ={
-        'students':students,
-    }
-    return render(request, 'students/edit.html', context)
-
-def update(request,pk):
-
-    name = request.POST.get('name')
-    age = request.POST.get('age')
-    students = Students.objects.get(pk=pk)
     
-    students.name = name
-    students.age = age
-    students.save()
+    if request.method == 'POST':
+        # update (def update)
+        name = request.POST.get('name')
+        age = request.POST.get('age')
+        
+    
+        students.name = name
+        students.age = age
+        students.save()
 
-    return redirect('/students/')
+        return redirect('/students/')
+
+    else:
+        # 수정하는 페이지 보여줌 
+        context ={
+            'students':students,
+        }
+        return render(request, 'students/edit.html', context)
+
+
+
+
+    
+
+# def update(request,pk):
+
+#     name = request.POST.get('name')
+#     age = request.POST.get('age')
+#     students = Students.objects.get(pk=pk)
+    
+#     students.name = name
+#     students.age = age
+#     students.save()
+
+#     return redirect('/students/')
